Refactor for conditional render on mobile
f

names
