+++ src/modules/dashboard/components/ColonyTotalFunds/ColonyTotalFunds.tsx
@@ -1,44 +1,20 @@
-import React, { useMemo, useState, useEffect } from 'react';
+import React from 'react';
 import { defineMessages, FormattedMessage } from 'react-intl';
 import { useMediaQuery } from 'react-responsive';
 import { ColonyVersion } from '@colony/colony-js';
 
-import Icon from '~core/Icon';
-import Link from '~core/Link';
-import Numeral from '~core/Numeral';
-import { MiniSpinnerLoader } from '~core/Preloaders';
-import {
-  Colony,
-  useTokenBalancesForDomainsQuery,
-  useLoggedInUser,
-} from '~data/index';
-import { Address } from '~types/index';
-import { getTokenDecimalsWithFallback } from '~utils/tokens';
-import { COLONY_TOTAL_BALANCE_DOMAIN_ID } from '~constants';
+import { Colony, useLoggedInUser } from '~data/index';
 import { checkIfNetworkIsAllowed } from '~utils/networks';
 
-import ColonyTotalFundsPopover from './ColonyTotalFundsPopover';
-
 import styles from './ColonyTotalFunds.css';
-import IconTooltip from '~core/IconTooltip';
 import { mobile } from '~utils/mediaQueries';
+import ColonyTotalFundsSelectedToken from './ColonyTotalFundsSelectedToken';
+import ColonyTotalFundsManageFunds from './ColonyTotalFundsManageFunds';
 
 const MSG = defineMessages({
   totalBalance: {
     id: 'dashboard.ColonyTotalFunds.totalBalance',
     defaultMessage: 'Colony total balance',
-  },
-  manageFundsLink: {
-    id: 'dashboard.ColonyTotalFunds.manageFundsLink',
-    defaultMessage: 'Manage Funds',
-  },
-  loadingData: {
-    id: 'dashboard.ColonyTotalFunds.loadingData',
-    defaultMessage: 'Loading token information...',
-  },
-  tokenSelect: {
-    id: 'dashboard.ColonyTotalFunds.tokenSelect',
-    defaultMessage: 'Select Tokens',
   },
 });
 
@@ -48,175 +24,35 @@ type Props = {
 
 const displayName = 'dashboard.ColonyTotalFunds';
 
-const ColonyTotalFunds = ({
-  colony: {
-    colonyAddress,
-    colonyName,
-    tokens: colonyTokens,
-    nativeTokenAddress,
-    version,
-    isNativeTokenLocked,
-  },
-}: Props) => {
+const ColonyTotalFunds = ({ colony: { version }, colony }: Props) => {
   const { networkId } = useLoggedInUser();
 
-  const [currentTokenAddress, setCurrentTokenAddress] = useState<Address>(
-    nativeTokenAddress,
-  );
-
   const isMobile = useMediaQuery({ query: mobile });
-
-  const {
-    data,
-    loading: isLoadingTokenBalances,
-  } = useTokenBalancesForDomainsQuery({
-    variables: {
-      colonyAddress,
-      domainIds: [COLONY_TOTAL_BALANCE_DOMAIN_ID],
-      tokenAddresses: colonyTokens.map(({ address }) => address),
-    },
-  });
-
-  useEffect(() => {
-    setCurrentTokenAddress(nativeTokenAddress);
-  }, [nativeTokenAddress]);
-
-  const currentToken = useMemo(() => {
-    if (data && data.tokens) {
-      return data.tokens.find(
-        ({ address: tokenAddress }) => tokenAddress === currentTokenAddress,
-      );
-    }
-    return undefined;
-  }, [data, currentTokenAddress]);
 
   const isSupportedColonyVersion =
     parseInt(version, 10) >= ColonyVersion.LightweightSpaceship;
   const isNetworkAllowed = checkIfNetworkIsAllowed(networkId);
 
-  if (!data || !currentToken || isLoadingTokenBalances) {
-    return (
-      <MiniSpinnerLoader
-        className={styles.main}
-        loadingText={MSG.loadingData}
-        titleTextValues={{ hasCounter: false }}
-      />
-    );
-  }
-
   return (
     <div className={styles.main}>
       {isMobile ? (
-        // Render Link outside of .totalBalanceCopy and render
-        // .totalBalanceCopy inside of .selectedToken on mobile
         <>
-          <div className={styles.selectedToken}>
-            <Numeral
-              className={styles.selectedTokenAmount}
-              unit={getTokenDecimalsWithFallback(currentToken.decimals)}
-              value={
-                currentToken.balances[COLONY_TOTAL_BALANCE_DOMAIN_ID].amount
-              }
-              data-test="colonyTotalFunds"
-            />
-            <ColonyTotalFundsPopover
-              tokens={data.tokens}
-              onSelectToken={setCurrentTokenAddress}
-              currentTokenAddress={currentTokenAddress}
-            >
-              <button className={styles.selectedTokenSymbol} type="button">
-                <span data-test="colonyTokenSymbol">{currentToken.symbol}</span>
-                {currentTokenAddress === nativeTokenAddress &&
-                  isNativeTokenLocked && (
-                    <IconTooltip
-                      icon="lock"
-                      tooltipText={{ id: 'tooltip.lockedToken' }}
-                      className={styles.tokenLockWrapper}
-                      appearance={{ size: 'large' }}
-                    />
-                  )}
-                <span className={styles.caretContainer}>
-                  <Icon
-                    className={styles.caretIcon}
-                    name="caret-down-small"
-                    title={MSG.tokenSelect}
-                  />
-                </span>
-              </button>
-            </ColonyTotalFundsPopover>
+          <ColonyTotalFundsSelectedToken colony={colony}>
             <div className={styles.totalBalanceCopy}>
               <FormattedMessage {...MSG.totalBalance} />
             </div>
-          </div>
+          </ColonyTotalFundsSelectedToken>
           {isSupportedColonyVersion && isNetworkAllowed && (
-            <Link
-              className={styles.manageFundsLink}
-              to={`/colony/${colonyName}/funds`}
-              data-test="manageFunds"
-            >
-              <Icon
-                className={styles.rightArrowDisplay}
-                name="arrow-right"
-                appearance={{ size: 'small' }}
-                title={MSG.manageFundsLink}
-              />
-              <FormattedMessage {...MSG.manageFundsLink} />
-            </Link>
+            <ColonyTotalFundsManageFunds colony={colony} />
           )}
         </>
       ) : (
         <>
-          <div className={styles.selectedToken}>
-            <Numeral
-              className={styles.selectedTokenAmount}
-              unit={getTokenDecimalsWithFallback(currentToken.decimals)}
-              value={
-                currentToken.balances[COLONY_TOTAL_BALANCE_DOMAIN_ID].amount
-              }
-              data-test="colonyTotalFunds"
-            />
-            <ColonyTotalFundsPopover
-              tokens={data.tokens}
-              onSelectToken={setCurrentTokenAddress}
-              currentTokenAddress={currentTokenAddress}
-            >
-              <button className={styles.selectedTokenSymbol} type="button">
-                <span data-test="colonyTokenSymbol">{currentToken.symbol}</span>
-                {currentTokenAddress === nativeTokenAddress &&
-                  isNativeTokenLocked && (
-                    <IconTooltip
-                      icon="lock"
-                      tooltipText={{ id: 'tooltip.lockedToken' }}
-                      className={styles.tokenLockWrapper}
-                      appearance={{ size: 'large' }}
-                    />
-                  )}
-                <span className={styles.caretContainer}>
-                  <Icon
-                    className={styles.caretIcon}
-                    name="caret-down-small"
-                    title={MSG.tokenSelect}
-                  />
-                </span>
-              </button>
-            </ColonyTotalFundsPopover>
-          </div>
+          <ColonyTotalFundsSelectedToken colony={colony} />
           <div className={styles.totalBalanceCopy}>
             <FormattedMessage {...MSG.totalBalance} />
             {isSupportedColonyVersion && isNetworkAllowed && (
-              <Link
-                className={styles.manageFundsLink}
-                to={`/colony/${colonyName}/funds`}
-                data-test="manageFunds"
-              >
-                <Icon
-                  className={styles.rightArrowDisplay}
-                  name="arrow-right"
-                  appearance={{ size: 'small' }}
-                  title={MSG.manageFundsLink}
-                />
-                <FormattedMessage {...MSG.manageFundsLink} />
-              </Link>
+              <ColonyTotalFundsManageFunds colony={colony} />
             )}
           </div>
         </>
